Renamed "GameArt" to CallAscii.java Renamed "HTTP" to AsciiArt.java AsciiArt will store all the Strings and ascii text, While CallAscii will inherit those methods for Main to call
