+++ src/io/trivia/CallAscii.java
@@ -2,50 +2,24 @@ package io.trivia;
 
 import java.lang.Thread;
 
-public class GameArt implements Runnable {
+public class CallAscii implements Runnable {
 
-	public GameArt() {
+
+	public CallAscii() {
 	}
 
 	@Override
 	public void run() {
 		Thread thread = new Thread();
-		GameArt art = new GameArt();
+		CallAscii art = new CallAscii();
 	}
+
 	/*
 	╭───────────╮
 	│loadFrame()│
 	╰───────────╯
 	*/
-	public void loadFrame() throws InterruptedException {
-
-
-
-
-		//
-		System.out.println("  _______ _                           _   _           ______                    _   _                   _ ");
-		Thread.sleep(400);
-		System.out.println(" |__   __| |                         | \\ | |         |  ____|                  | | (_)                 | |");
-		Thread.sleep(375);
-		System.out.println("    | |  | |__  _ __ _____      __   |  \\| | ___     | |__  __  _____ ___ _ __ | |_ _  ___  _ __  ___  | |");
-		Thread.sleep(350);
-		System.out.println("    | |  | '_ \\| '__/ _ \\ \\ /\\ / /   | . ` |/ _ \\    |  __| \\ \\/ / __/ _ \\ '_ \\| __| |/ _ \\| '_ \\/ __| | |");
-		Thread.sleep(325);
-		System.out.println("    | |  | | | | | | (_) \\ V  V /    | |\\  | (_) |   | |____ >  < (_|  __/ |_) | |_| | (_) | | | \\__ \\ |_|");
-		Thread.sleep(300);
-		System.out.println("    |_|  |_| |_|_|  \\___/ \\_/\\_/     |_| \\_|\\___/    |______/_/\\_\\___\\___| .__/ \\__|_|\\___/|_| |_|___/ (_)");
-		Thread.sleep(275);
-		System.out.println("                                                                         | |                              ");
-		Thread.sleep(250);
-		System.out.println("                                                                         |_|                              ");
-	}
-
-
-
-
-
-
-
+	public void loadFrame() throws InterruptedException {}
 
 
 
@@ -55,7 +29,7 @@ public class GameArt implements Runnable {
 	│loadTitle()│
 	╰───────────╯
 	*/
-	public void loadTitle() throws InterruptedException {
+/*	public void loadTitle() throws InterruptedException {
 		System.out.println("  _______ _                           _   _           ______                    _   _                   _ ");
 		Thread.sleep(400);
 		System.out.println(" |__   __| |                         | \\ | |         |  ____|                  | | (_)                 | |");
@@ -71,31 +45,14 @@ public class GameArt implements Runnable {
 		System.out.println("                                                                         | |                              ");
 		Thread.sleep(250);
 		System.out.println("                                                                         |_|                              ");
-	}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+	}*/
 }
+
+
+
+
+
+
